Add more paths to clean7days script

+++ clean7days.sh
@@ -3,7 +3,11 @@
 # Selects downloads directory
 [ -f ${HOME}/.config/user-dirs.dirs ] && . ${HOME}/.config/user-dirs.dirs
 
-dirs="${XDG_DOWNLOAD_DIR} ${XDG_CACHE_DIR}"
+dirs="${XDG_DOWNLOAD_DIR}
+${XDG_CACHE_DIR}
+/storage/emulated/0/Download
+/storage/emulated/0/DCIM
+${HOME}/.var/app/org.telegram.desktop/data/TelegramDesktop/tdata/temp_data/"
 
 [ -z "$dirs" ] && [ -z "$1" ] && exit
 
@@ -17,17 +21,17 @@ dirs="${XDG_DOWNLOAD_DIR} ${XDG_CACHE_DIR}"
 
 for dir in $dirs
 do
-	(
-	# Step 1: Remove files older than 7 days
-	find ${dir}/ -mindepth 1 -type f -mtime +${days} -exec rm -v {} \;
-	# Step 2: Remove all symlinks
-	find ${dir}/ -mindepth 1 -type l -exec rm -v {} \;
-	# Step 3: Delete empty folders recursively
-	find ${dir}/ -mindepth 1 -type d -depth -empty -exec rmdir -v {} \;
-	# Notify when done
-	notify-send "${dir} cleanup" "Remove files older than ${days} days"
-	) &
+	if [ -d "$dir" ]
+	then
+		# Step 1: Remove files older than 7 days
+		find ${dir}/ -mindepth 1 -type f -mtime +${days} -exec rm -v {} \;
+		# Step 2: Remove all symlinks
+		find ${dir}/ -mindepth 1 -type l -exec rm -v {} \;
+		# Step 3: Delete empty folders recursively
+		find ${dir}/ -mindepth 1 -depth -type d  -empty -exec rmdir -v {} \;
+		# Notify when done
+		notify-send "${dir} cleanup" "Remove files older than ${days} days"
+	fi
 done
-wait
 
 exit
